Only allow integer indices. The level of meta-programming in mul_leftright.jl got dialed up to 11.

+++ src/mul_leftright.jl
@@ -148,44 +148,44 @@ end
     l
 end
 
-@inline function mul_left!(r::PauliOperator, l::Tableau, i; phases::Val{B}=Val(true)) where B
+@inline function mul_left!(r::PauliOperator, l::Tableau, i::Integer; phases::Val{B}=Val(true)) where B
     s = mul_left!(r.xz, (@view l.xzs[:,i]), phases=phases)
     B && (r.phase[] = (s+r.phase[]+l.phases[i])&0x3)
     r
 end
 
-@inline mul_left!(r::PauliOperator, l::Stabilizer, i; phases::Val{B}=Val(true)) where B = mul_left!(r, tab(l), i; phases=phases)
+@inline mul_left!(r::PauliOperator, l::Stabilizer, i::Integer; phases::Val{B}=Val(true)) where B = mul_left!(r, tab(l), i; phases=phases)
 
-@inline function mul_right!(l::PauliOperator, r::Tableau, i; phases::Val{B}=Val(true)) where B
+@inline function mul_right!(l::PauliOperator, r::Tableau, i::Integer; phases::Val{B}=Val(true)) where B
     s = mul_right!(l.xz, (@view r.xzs[:,i]), phases=phases)
     B && (l.phase[] = (s+l.phase[]+r.phases[i])&0x3)
     l
 end
 
-@inline mul_right!(l::PauliOperator, r::Stabilizer, i; phases::Val{B}=Val(true)) where B = mul_right!(l, tab(r), i; phases=phases)
+@inline mul_right!(l::PauliOperator, r::Stabilizer, i::Integer; phases::Val{B}=Val(true)) where B = mul_right!(l, tab(r), i; phases=phases)
 
 ##############################
 # On Tableaux
 ##############################
 
-@inline function mul_left!(s::Tableau, m, t::Tableau, i; phases::Val{B}=Val(true)) where B
+@inline function mul_left!(s::Tableau, m::Integer, t::Tableau, i::Integer; phases::Val{B}=Val(true)) where B
     extra_phase = mul_left!((@view s.xzs[:,m]), (@view t.xzs[:,i]); phases=phases)
     B && (s.phases[m] = (extra_phase+s.phases[m]+s.phases[i])&0x3)
     s
 end
 
-@inline function mul_left!(s::Tableau, m, i; phases::Val{B}=Val(true)) where B
+@inline function mul_left!(s::Tableau, m::Integer, i::Integer; phases::Val{B}=Val(true)) where B
     extra_phase = mul_left!((@view s.xzs[:,m]), (@view s.xzs[:,i]); phases=phases)
     B && (s.phases[m] = (extra_phase+s.phases[m]+s.phases[i])&0x3)
     s
 end
 
-@inline function mul_left!(s::Stabilizer, m, i; phases::Val{B}=Val(true)) where B
+@inline function mul_left!(s::Stabilizer, m::Integer, i::Integer; phases::Val{B}=Val(true)) where B
     mul_left!(tab(s), m, i; phases=phases)
     s
 end
 
-@inline function mul_left!(s::Destabilizer, i, j; phases::Val{B}=Val(true)) where B
+@inline function mul_left!(s::Destabilizer, i::Integer, j::Integer; phases::Val{B}=Val(true)) where B
     t = tab(s)
     mul_left!(t, j, i; phases=Val(false)) # Indices are flipped to preserve commutation constraints
     n = size(t,1)÷2
@@ -193,12 +193,12 @@ end
     s
 end
 
-@inline function mul_left!(s::MixedStabilizer, i, j; phases::Val{B}=Val(true)) where B
+@inline function mul_left!(s::MixedStabilizer, i::Integer, j::Integer; phases::Val{B}=Val(true)) where B
     mul_left!(tab(s), i, j; phases=phases)
     s
 end
 
-@inline function mul_left!(s::MixedDestabilizer, i, j; phases::Val{B}=Val(true)) where B
+@inline function mul_left!(s::MixedDestabilizer, i::Integer, j::Integer; phases::Val{B}=Val(true)) where B
     t = tab(s)
     mul_left!(t, j, i; phases=Val(false)) # Indices are flipped to preserve commutation constraints
     n = nqubits(t)
